requirements.txt
Django==1.5.1
MySQL-python==1.2.3
argparse==1.2.1
boto==2.9.2
distribute==0.6.36
django-debug-toolbar==0.9.4
django-social-auth==0.7.23
django-storages==1.1.8
django-tastypie==0.9.14
gunicorn==0.17.4
httplib2==0.8
mimeparse==0.1.3
oauth2==1.5.211
python-dateutil==2.1
python-openid==2.2.5
six==1.3.0
wsgiref==0.1.2
